refactor 폴더 구조 재수정'
