+++ src/refactoring/components/cart/CartItemDetails.tsx
@@ -1,4 +1,4 @@
-import { CartItem } from "../../../../types";
+import { CartItem } from "../../../types";
 
 interface CartItemDetailsProps {
   item: CartItem;
