+++ src/refactoring/pages/AdminPage.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Coupon, Discount, Product } from "../../../../types.ts";
+import { Coupon, Discount, Product } from "../../types.ts";
 
 interface Props {
   products: Product[];
